Extracted merged_directories
