+++ conans/util/files.py
@@ -363,3 +363,59 @@ def exception_message_safe(exc):
         return str(exc)
     except Exception:
         return decode_text(repr(exc))
+
+
+def merge_directories(src, dst, excluded=None):
+    src = os.path.normpath(src)
+    dst = os.path.normpath(dst)
+    excluded = excluded or []
+    excluded = [os.path.normpath(entry) for entry in excluded]
+
+    def is_excluded(origin_path):
+        if origin_path == dst:
+            return True
+        rel_path = os.path.normpath(os.path.relpath(origin_path, src))
+        if rel_path in excluded:
+            return True
+        return False
+
+    def link_to_rel(pointer_src):
+        linkto = os.readlink(pointer_src)
+        if not os.path.isabs(linkto):
+            linkto = os.path.join(os.path.dirname(pointer_src), linkto)
+
+        # Check if it is outside the sources
+        out_of_source = os.path.relpath(linkto, os.path.realpath(src)).startswith(".")
+        if out_of_source:
+            # May warn about out of sources symlink
+            return
+
+        # Create the symlink
+        linkto_rel = os.path.relpath(linkto, os.path.dirname(pointer_src))
+        pointer_dst = os.path.normpath(os.path.join(dst, os.path.relpath(pointer_src, src)))
+        os.symlink(linkto_rel, pointer_dst)
+
+    for src_dir, dirs, files in walk(src, followlinks=True):
+        if is_excluded(src_dir):
+            dirs[:] = []
+            continue
+
+        if os.path.islink(src_dir):
+            link_to_rel(src_dir)
+            dirs[:] = []  # Do not enter subdirectories
+            continue
+
+        # Overwriting the dirs will prevents walk to get into them
+        files[:] = [d for d in files if not is_excluded(os.path.join(src_dir, d))]
+
+        dst_dir = os.path.normpath(os.path.join(dst, os.path.relpath(src_dir, src)))
+        if not os.path.exists(dst_dir):
+            os.makedirs(dst_dir)
+        for file_ in files:
+            src_file = os.path.join(src_dir, file_)
+            dst_file = os.path.join(dst_dir, file_)
+            if os.path.islink(src_file):
+                link_to_rel(src_file)
+            else:
+                shutil.copy2(src_file, dst_file)
+
